+++ src/global.php
@@ -8,18 +8,18 @@ if (!function_exists('json5_decode')) {
      * The parameters exactly match PHP's json_decode() function - see
      * http://php.net/manual/en/function.json-decode.php for more information.
      *
-     * @param string $source      The JSON string being decoded.
+     * @param string $json        The JSON string being decoded.
      * @param bool   $associative When TRUE, returned objects will be converted into associative arrays.
      * @param int    $depth       User specified recursion depth.
-     * @param int    $options     Bitmask of JSON decode options.
+     * @param int    $flags       Bitmask of JSON decode options.
      *
      * @throws \ColinODell\Json5\SyntaxError if the JSON encoded string could not be parsed.
      *
      * @return mixed
      */
-    function json5_decode(string $source, ?bool $associative = false, int $depth = 512, int $options = 0)
+    function json5_decode(string $json, ?bool $associative = false, int $depth = 512, int $flags = 0)
     {
-        return \ColinODell\Json5\Json5Decoder::decode($source, $associative, $depth, $options);
+        return \ColinODell\Json5\Json5Decoder::decode($json, $associative, $depth, $flags);
     }
 }